// web/src/api/generated/models/AccountView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AccountView
 */
export interface AccountView {
    /**
     * 
     * @type {string}
     * @memberof AccountView
     */
    displayName: string;
    /**
     * 
     * @type {string}
     * @memberof AccountView
     */
    id: string;
}

/**
 * Check if a given object implements the AccountView interface.
 */
export function instanceOfAccountView(value: object): value is AccountView {
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function AccountViewFromJSON(json: any): AccountView {
    return AccountViewFromJSONTyped(json, false);
}

export function AccountViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountView {
    if (json == null) {
        return json;
    }
    return {
        
        'displayName': json['displayName'],
        'id': json['id'],
    };
}

export function AccountViewToJSON(json: any): AccountView {
    return AccountViewToJSONTyped(json, false);
}

export function AccountViewToJSONTyped(value?: AccountView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'displayName': value['displayName'],
        'id': value['id'],
    };
}